+++ src/test/java/mod/tjt01/lapislibtest/block/LapisLibTestBlocks.java
@@ -31,6 +31,8 @@ public class LapisLibTestBlocks {
 
     public static final RegistryObject<Block> OPTIONAL_BLOCK = makeBlock("optional_block", () -> new OptionalBlock(BlockBehaviour.Properties.of(Material.STONE), () -> false), CreativeModeTab.TAB_BUILDING_BLOCKS);
 
+    public static final RegistryObject<Block> CRAFTING_BLOCK = makeBlock("test_crafting_block", () -> new TestCraftingBlock(BlockBehaviour.Properties.of(Material.STONE)), CreativeModeTab.TAB_DECORATIONS);
+
     public static void register(IEventBus bus) {
         BLOCKS.register(bus);
     }