+++ src/test/java/mod/tjt01/lapislibtest/data/gen/DataGenerators.java
@@ -1,6 +1,7 @@
 package mod.tjt01.lapislibtest.data.gen;
 
 import mod.tjt01.lapislibtest.LapisLibTest;
+import mod.tjt01.lapislibtest.data.gen.asset.BlockStates;
 import net.minecraft.data.DataGenerator;
 import net.minecraftforge.common.data.ExistingFileHelper;
 import net.minecraftforge.eventbus.api.SubscribeEvent;
@@ -16,5 +17,6 @@ public class DataGenerators {
         generator.addProvider(new TestRecipes(generator));
         generator.addProvider(new LootModifiers(generator));
         generator.addProvider(new Lang(generator));
+        generator.addProvider(new BlockStates(generator, helper));
     }
 }