+++ src/test/java/mod/tjt01/lapislibtest/data/recipe/package-info.java
@@ -1,0 +1,7 @@
+@MethodsReturnNonnullByDefault @ParametersAreNonnullByDefault @FieldsAreNonnullByDefault
+package mod.tjt01.lapislibtest.data.recipe;
+
+import net.minecraft.FieldsAreNonnullByDefault;
+import net.minecraft.MethodsReturnNonnullByDefault;
+
+import javax.annotation.ParametersAreNonnullByDefault;